refactor: PointHistory infrastructure 구성
