+++ src/main/java/com/example/surveyapp/domain/point/infrastructure/PointHistoryJpaRepository.java
@@ -1,11 +1,11 @@
-package com.example.surveyapp.domain.point.domain.repository;
+package com.example.surveyapp.domain.point.infrastructure;
 
 import com.example.surveyapp.domain.point.domain.model.entity.PointHistory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PointHistoryRepository {
+public interface PointHistoryJpaRepository extends JpaRepository<PointHistory, Long> {
     Page<PointHistory> findPointHistoryByUserId (Long userId, Pageable pageable);
 
     PointHistory save(PointHistory history);
